Move default settings to `src/main/settings.ts`
