+++ package.json
@@ -91,6 +91,7 @@
     "devDependencies": {
         "@playwright/test": "^1.16.3",
         "@types/chmodr": "1.0.0",
+        "@types/lodash.merge": "4.6.7",
         "@types/mustache": "4.2.1",
         "@types/node": "14.17.17",
         "@types/targz": "1.0.1",
@@ -114,6 +115,7 @@
         "electron-store": "8.0.1",
         "electron-updater": "4.3.1",
         "fs-extra": "9.0.0",
+        "lodash.merge": "4.6.2",
         "mustache": "4.0.1",
         "serialport": "8.0.8",
         "shasum": "1.0.2",
